[eds_move_down] move xds balancer eds wrapper one level down
