+++ xds/internal/balancer/edsbalancer/config.go
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package balancer
+package edsbalancer
 
 import (
 	"encoding/json"
