+++ xds/internal/balancer/edsbalancer/xds_client_test.go
@@ -16,7 +16,7 @@
  *
  */
 
-package balancer
+package edsbalancer
 
 import (
 	"errors"
