+++ xds/internal/balancer/edsbalancer/xds_client_wrapper.go
@@ -16,7 +16,7 @@
  *
  */
 
-package balancer
+package edsbalancer
 
 import (
 	"google.golang.org/grpc"
