+++ xds/internal/balancer/edsbalancer/xds_old.go
@@ -13,9 +13,10 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-package balancer
+package edsbalancer
 
 import "google.golang.org/grpc/balancer"
